renamed DiktatRuleSetFactory back to DiktatRuleSetProvider
