+++ diktat-rules/src/main/kotlin/org/cqfn/diktat/ktlint/KtLintRuleSetProviderWrapper.kt
@@ -1,7 +1,7 @@
 package org.cqfn.diktat.ktlint
 
 import org.cqfn.diktat.ktlint.KtLintRuleSetWrapper.Companion.toKtLint
-import org.cqfn.diktat.ruleset.rules.DiktatRuleSetFactory
+import org.cqfn.diktat.ruleset.rules.DiktatRuleSetProvider
 import com.pinterest.ktlint.core.RuleSet
 import com.pinterest.ktlint.core.RuleSetProvider
 
@@ -9,14 +9,14 @@ import com.pinterest.ktlint.core.RuleSetProvider
  * This is a wrapper around __KtLint__'s [RuleSetProvider].
  */
 class KtLintRuleSetProviderWrapper private constructor(
-    private val diktatRuleSetFactory: DiktatRuleSetFactory,
+    private val diktatRuleSetFactory: DiktatRuleSetProvider,
 ) : RuleSetProvider {
     override fun get(): RuleSet = diktatRuleSetFactory().toKtLint()
 
     companion object {
         /**
-         * @return __KtLint__'s [RuleSetProvider] created from [DiktatRuleSetFactory]
+         * @return __KtLint__'s [RuleSetProvider] created from [DiktatRuleSetProvider]
          */
-        fun DiktatRuleSetFactory.toKtLint(): RuleSetProvider = KtLintRuleSetProviderWrapper(this)
+        fun DiktatRuleSetProvider.toKtLint(): RuleSetProvider = KtLintRuleSetProviderWrapper(this)
     }
 }
