+++ diktat-rules/src/main/resources/META-INF/services/com.pinterest.ktlint.core.RuleSetProvider
@@ -1,1 +1,1 @@
-org.cqfn.diktat.ruleset.rules.DiktatRuleSetProvider
+org.cqfn.diktat.ruleset.rules.DiktatRuleSetProviderSpi
